+++ lib/core/KuzzleProxy.js
@@ -1,7 +1,7 @@
 'use strict';
 
 let
-  debug = require('../debug')('kuzzle-proxy:plugins'),
+  debug = require('../kuzzleDebug')('kuzzle-proxy:plugins'),
   Broker = require('../service/Broker'),
   ClientConnectionStore = require('../store/ClientConnection'),
   Context = require('./Context'),
