+++ lib/service/Backend.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const
-  debug = require('../debug')('kuzzle-proxy:backend'),
+  debug = require('../kuzzleDebug')('kuzzle-proxy:backend'),
   PendingRequest = require('../store/PendingRequest'),
   util = require('util'),
   InternalError = require('kuzzle-common-objects').errors.InternalError;
