+++ lib/service/Broker.js
@@ -1,7 +1,7 @@
 'use strict';
 
 let
-  debug = require('../debug')('kuzzle-proxy:broker'),
+  debug = require('../kuzzleDebug')('kuzzle-proxy:broker'),
   fs = require('fs'),
   http = require('http'),
   net = require('net'),
