+++ lib/service/protocol/SocketIo.js
@@ -2,7 +2,7 @@
 
 const
   _protocol = 'socketio',
-  debug = require('../../debug')('kuzzle-proxy:socketio-protocol'),
+  debug = require('../../kuzzleDebug')('kuzzle-proxy:socketio-protocol'),
   ClientConnection = require('../../core/clientConnection'),
   Request = require('kuzzle-common-objects').Request;
