+++ lib/service/protocol/Websocket.js
@@ -2,7 +2,7 @@
 
 const
   _protocol = 'websocket',
-  debug = require('../../debug')('kuzzle-proxy:websocket-protocol'),
+  debug = require('../../kuzzleDebug')('kuzzle-proxy:websocket-protocol'),
   WebSocketServer = require('ws').Server,
   ClientConnection = require('../../core/clientConnection'),
   Request = require('kuzzle-common-objects').Request,
